Log error in Rails
When developing in Rails, and even when running in production, I would
like to be able to see backtraces and errors in my logs.

This snippet of code is taken from the Rails source code, where it
appears in multiple non-reusable private methods. It will give the
familiar Rails feel when looking at one's log.